+++ src/Pho/Framework/Handlers/Set.php
@@ -12,12 +12,21 @@
 namespace Pho\Framework\Handlers;
 
 use Pho\Framework\ParticleInterface;
+use Pho\Framework\FieldHelper;
 use Pho\Framework\Cargo\FieldsCargo;
 use Pho\Framework\Exceptions\InvalidEdgeHeadTypeException;
-use Webmozart\Assert\Assert;
 
 /**
  * Setter Handler
+ * 
+ * A setter may have two forms;
+ * * ```setSomething()``` which would be used for Fields, where Something is a Field.
+ * * ```something()``` which would be used for outgoing edges.
+ * 
+ * In this class we don't modify $name arguments (unlike what we do in Form, Get and Has classes)
+ * because we expect the method to be called in the proper format, and:
+ * * for Fields, they are already stored in the particle in upper-camelized format.
+ * * for outgoing edges, they are already stored in the particle in camlized format.
  * 
  * @author Emre Sokullu <emre@phonetworks.org>
  */
@@ -36,8 +45,11 @@ class Set implements HandlerInterface
         array $args
         ) /*:  \Pho\Lib\Graph\EntityInterface*/
     {
+        // don't camelize or upper-camelize $name yet because it is supposed to
+        // come in upper-camlized format, and that's how they are stored by 
+        // Loaders\FieldLoader.php
         if( Utils::fieldExists($pack["fields"], substr($name, 3)) ) {
-            return static::field($particle, $pack["fields"], substr($name, 3), $args[0]);
+            return static::field($particle, $pack["fields"], substr($name, 3), $args);
         }
         $check = false;
         foreach($pack["out"]->setter_label_settable_pairs[$name] as $settable) {
@@ -46,7 +58,7 @@ class Set implements HandlerInterface
         if(!$check) { 
             throw new InvalidEdgeHeadTypeException($args[0], $pack["out"]->setter_label_settable_pairs[$name]);
         }
-        $edge = new $pack["out"]->setter_classes[$name]($particle, $args[0]);
+        $edge = new $pack["out"]->setter_classes[$name]($particle, array_shift($args), null, $args);
         return $edge->return();
     }
 
@@ -56,7 +68,7 @@ class Set implements HandlerInterface
      * @param ParticleInterface $particle
      * @param FieldsCargo $cargo
      * @param string $name Field name
-     * @param array $args Field argument
+     * @param array $args Field argument; [0] mixed, the argument. [1] ?bool defer_persist
      * 
      * @return void
      * 
@@ -66,69 +78,46 @@ class Set implements HandlerInterface
         ParticleInterface $particle,
         FieldsCargo $cargo,
         string $name,
-        /*mixed*/ $value
+        array $value 
         ): void
     {
+        $defer_persist = false;
+        $defer_persist = ( isset($value[1]) && $value[1] == true );
+        $value = $value[0];
         $name = Utils::findFieldName($cargo, $name);
-        if(isset($cargo->fields[$name]["constraints"])) {
-                static::probeField($cargo->fields[$name]["constraints"], $value);
-        }
-        $particle->attributes()->$name = static::applyDirectives($value, $cargo->fields[$name]);
+        $field_helper = new FieldHelper($value, $cargo->fields[$name]);
+        $field_helper->probe();
+        $value = $field_helper->process(); 
+        static::saveField($particle, $name, $value, $defer_persist, $field_helper);
     }
-
+    
     /**
-     * Applies directives to the value to return
+     * Helper method to save the field
      * 
-     * @param mixed $value The value to check.
-     * @param array $directives Particle directives.
-     * 
-     * @return mixed 
-     */
-    protected static function applyDirectives(/*mixed*/ $value, array $field_settings) /*: mixed*/
-    {
-        if(!isset($field_settings["directives"]))
-            return $value;
-        $directives = $field_settings["directives"];
-        $isDirectiveEnabled = function(string $param) use($directives): bool
-        {
-            return (isset($directives[$param]) && $directives[$param]);
-        };
-        if($isDirectiveEnabled("md5")) {
-            return md5($value);
-        }
-        return $value;
-    }
-
-    /**
-     * Checks if the field meets the requirements of the constraints in the 
-     * particle's  FIELDS constant.
-     * 
-     * @param array $constraints
-     * @param [type] $field_value
+     * This method exists to help extending the "field" method without overriding it.
+     *
+     * @param ParticleInterface $particle
+     * @param string $field_name Field name
+     * @param mixed $field_value Field value
+     * @param bool $defer_persist Whether to defer persistence
+     * @param FieldHelper $helper an object full of information re: this field.
      * 
      * @return void
-     * 
+     *
      * @throws \InvalidArgumentException thrown when there argument does not meet the constraints.
      */
-    protected static function probeField(array $constraints, $field_value): void
+    protected static function saveField(
+        ParticleInterface $particle, 
+        string $field_name, 
+        /*mixed*/ $field_value, 
+        bool $defer_persist, 
+        FieldHelper $helper): void
     {
-        foreach($constraints as $constraint=>$constraint_val) {
-            if(is_null($constraint_val))
-                continue;
-            switch($constraint) {
-                case "minLength":
-                case "maxLength":
-                case "greaterThan":
-                case "lessThan":
-                    Assert::$constraint($field_value, $constraint_val);
-                    break;
-                case "uuid":
-                    Assert::$constraint($field_value);
-                    break;
-                case "regex":
-                    Assert::$constraint($field_value, "/".addslashes($constraint_val)."/");
-                    break;
-             } 
+
+        if(!$defer_persist) {
+            $particle->attributes()->$field_name = $field_value;
+            return;
         }
+        $particle->attributes()->quietSet($field_name, $field_value);
     }
 }