+++ tests/Pho/Framework/FieldsTest.php
@@ -53,6 +53,68 @@ class ConstraintedTestObject extends Object
     ];
 }
 
+class ExtraConstraintedTestObject extends Object
+{
+    const FIELDS = [
+        "my_date" => [
+            "constraints" => [
+                "minLength" => null,
+                "maxLength" => null,
+                "uuid" => null,
+                "regex" => null,
+                "greaterThan" => null,
+                "lessThan" => null,
+                "dateAfter" => "1/15/1983",
+                "dateBefore" => "1/15/2017",
+            ],
+            "directives" => [
+                "md5" => false,
+                "now" => false,
+                "default" => "",
+            ]
+        ],
+        "my_int" => [
+            "constraints" => [
+                "minLength" => null,
+                "maxLength" => null,
+                "uuid" => null,
+                "regex" => null,
+                "greaterThan" => null,
+                "lessThan" => null,
+                "lessThan" => 40,
+                "greaterThank" => 10,
+            ],
+            "directives" => [
+                "md5" => false,
+                "now" => false,
+                "default" => "",
+            ]
+        ]
+    ];
+}
+
+class sha1Object extends Object
+{
+    const FIELDS = [
+        "my_field" => [
+            "constraints" => [
+                "minLength" => 6,
+                "maxLength" => null,
+                "uuid" => null,
+                "regex" => null,
+                "greaterThan" => null,
+                "lessThan" => null,
+            ],
+            "directives" => [
+                "sha1" => true,
+                "md5" => false,
+                "now" => false,
+                "default" => "",
+            ]
+        ]
+    ];
+}
+
 class FieldsTest extends \PHPUnit\Framework\TestCase 
 {
     private $space, $actor;
@@ -97,7 +159,7 @@ class FieldsTest extends \PHPUnit\Framework\TestCase
                 ]
             ];
         };
-        $uuid = \Pho\Lib\Graph\ID::generate();
+        $uuid = \Pho\Lib\Graph\ID::generate($obj);
         $obj->setMyField($uuid);
         $this->assertEquals($uuid, $obj->getMyField());
     }
@@ -107,12 +169,12 @@ class FieldsTest extends \PHPUnit\Framework\TestCase
             const FIELDS = [
                 "myField" => [
                     "constraints" => [
-                        "uuid" => true
+                        "id" => true
                     ]
                 ]
             ];
         };
-        $this->expectException("\InvalidArgumentException");
+        $this->expectException(\InvalidArgumentException::class);
         $obj->setMyField("not_uuid");
     }
 
@@ -145,6 +207,66 @@ class FieldsTest extends \PHPUnit\Framework\TestCase
         $obj->setMyField("will_not_match_regexp");
     }
 
+    public function testFormatConstraintsNegative() {
+        $obj = new class($this->actor, $this->space) extends Object {
+            const FIELDS = [
+                "MyField" => [
+                    "constraints" => [
+                        "format" => "email"
+                    ]
+                ]
+            ];
+        };
+        $this->expectException("\InvalidArgumentException");
+        $obj->setMyField("not_an_email!");
+    }
+
+    public function testDateConstraints() {
+        $obj = new class($this->actor, $this->space) extends Object {
+            const FIELDS = [
+                "MyField" => [
+                    "constraints" => [
+                        "dateBefore" => "01/20/2018",
+                        "dateAfter" => "01/15/2018",
+                    ]
+                ]
+            ];
+        };
+        //eval(\Psy\sh());
+        $this->expectException("\InvalidArgumentException");
+        $obj->setMyField("01/20/2018");
+    }
+
+    public function testDateConstraintsPositive() {
+        $obj = new class($this->actor, $this->space) extends Object {
+            const FIELDS = [
+                "MyField" => [
+                    "constraints" => [
+                        "dateBefore" => "01/20/2018",
+                        "dateAfter" => "01/15/2018",
+                    ]
+                ]
+            ];
+        };
+        $obj->setMyField("01/18/2018");
+        $this->assertEquals("01/18/2018", $obj->getMyField());
+    }
+
+    public function testFormatConstraintsPositive() {
+        $obj = new class($this->actor, $this->space) extends Object {
+            const FIELDS = [
+                "MyField" => [
+                    "constraints" => [
+                        "format" => "email"
+                    ]
+                ]
+            ];
+        };
+        $email = "emre@groups-inc.com";
+        $obj->setMyField($email);
+        $this->assertEquals($email, $obj->getMyField());
+    }
+
     public function testJsonFields() {
         $obj = new class($this->actor, $this->space)  extends Object {
             const FIELDS = '{"MyField":{"constraints":{"regex":"^A[0-9]+1$"}}}';
@@ -152,6 +274,79 @@ class FieldsTest extends \PHPUnit\Framework\TestCase
         $field_val = "A883841";
         $obj->setMyField($field_val);
         $this->assertEquals($field_val, $obj->getMyField());
+    }
+
+    public function testEdgeFields() {
+        $ref = 0;
+        $another_actor = new Actor($this->space);
+        $obj = new class($this->actor, $another_actor)  extends AbstractEdge {
+            const FIELDS = '{"my_field":{"constraints":{"regex":"^A[0-9]+1$"}}}';
+        };
+        $obj->on("modified", function() use(&$ref) {
+            $ref++;
+        });
+        $field_val1 = "A883841";
+        $field_val2 = "A4896541";
+        $obj->setMyField($field_val1);
+        $this->assertEquals($field_val1, $obj->getMyField());
+        $this->assertEquals(1, $ref); // signal
+        $obj->setMyField($field_val2, true);
+        $this->assertEquals($field_val2, $obj->getMyField());
+        $this->assertEquals(1, $ref); // no signal
+    }
+
+    public function testEdgeFieldsAutoPilot() {
+        $field_val1 = "A883841";
+        $past_time = 1502224636;
+        $another_actor = new Actor($this->space);
+        $obj = new class($this->actor, $another_actor, null, $field_val1)  extends AbstractEdge {
+            const FIELDS = '{"my_field":{"constraints":{"regex":"^A[0-9]+1$"}},"created_at":{"directives":{"now":true
+}},"with_default":{"directives":{"default":"defne"}}}'; 
+            // '{"my_field":{"constraints":{"regex":"^A[0-9]+1$"}},"created_at":{"directives":{"now":true}}}';
+        };
+        //eval(\Psy\sh());
+        $this->assertGreaterThan($past_time, $obj->getCreatedAt());
+        $this->assertEquals("defne", $obj->getWithDefault());
+    }
+
+    public function testSha1() {
+        
+        $obj = new sha1Object($this->actor, $this->space);
+        $obj->setMyField("emre12");
+        //$node_expected_to_be_identical = $this->space->get($node->id());
+        $this->assertEquals(sha1("emre12"), $obj->getMyField());
+    }
+
+    /**
+     * May be too similar to:
+     * @see testDateConstraintsPositive
+     */
+    public function testDateConstraintsPositive_withMultipleFields() {
+        $obj = new ExtraConstraintedTestObject($this->actor, $this->space);
+        $obj->setMyDate("10/10/2010");
+        $this->assertTrue(true);
+    }
+
+    /**
+     * May be too similar to:
+     * @see testDateConstraintsNegative
+     */
+    public function testDateConstraintsNegative_withMultipleFields() {
+        $obj = new ExtraConstraintedTestObject($this->actor, $this->space);
+        $this->expectException(\InvalidArgumentException::class);
+        $obj->setMyDate("10/10/1977");
+    }
+
+    public function testIntConstraintsPositive() {
+        $obj = new ExtraConstraintedTestObject($this->actor, $this->space);
+        $obj->setMyInt(20);
+        $this->assertTrue(true);
+    }
+
+    public function testIntConstraintsNegative() {
+        $obj = new ExtraConstraintedTestObject($this->actor, $this->space);
+        $this->expectException(\InvalidArgumentException::class);
+        $obj->setMyInt(1000);
     }
 
 }